cv-service/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
opencv-python==4.9.0.80
ultralytics==8.1.0
numpy>=1.26.0,<2.0.0
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
